fix: 쿠키 grade수정 쿠키 댓글수 -> 평점으로 UI 변경

+++ src/main/java/ovenbreak/ckrun/domain/cookies/Cookies.java
@@ -20,6 +20,6 @@ public class Cookies {
     private int startHealth;
     private int endHealth;
     private LocalDate rdate;
-    private int grade;
+    private float grade;
     private int comment;
 }
